refactor: seperate ingestor classes from the quote model class
